Buttons graying out bugfix (#24)
* Attempt to reduce memory usage, especially with canvas

Attempt to free up more ram, especially as the weekly art leaderboard seems to hog up a lot of ram.

* Fix buttons graying out

Fixes buttons graying out when doing an action on a verification, likely happening because of the first editReply due to some delay executing after the deny has already been handled, and thus turning gray instead of showing the "completed" form.