change how temp images are saved in S3
Temp images now are saved in a seperate directory called temp instead of with a _temp suffix. This might make it a bit easier in the future as it's better seperated. (for example to delete all temp images from a server at once etc.)

Still is kind of a mixed bag with older code that was for local image saving. This still needs to be cleaned.