update buttons&commands for multiple applications
Updated all buttons under applications to now use applications instead of the ServerConfig. Within the interaction customid the first context is the appname, which is used to identify the application the button is on.

Also changed seperate slash commands that modify serverconfig to now use applications. This changes nothing for the user except when there are multiple applications, then the user needs to supply the (autofill) application they want to use; for example with verifying and denying commands or toggling threads.

questionid model/table in database also has new column with the name of the application.

Also really shouldn't be in this branch, but fixed fonts in the artleaderboard by supplying them directly instead of relying on global fonts.

*note*: Like last commit, still is kind of a mixed bag with older code that was for local image saving.